+++ examples/resources/hrobot_rdns/resource.tf
@@ -1,0 +1,4 @@
+resource "hrobot_rdns" "example" {
+  ip  = "203.0.113.1"
+  ptr = "host.example.com"
+}