+++ examples/resources/hrobot_ssh_key/resource.tf
@@ -1,0 +1,4 @@
+resource "hrobot_ssh_key" "example" {
+  name       = "deploy"
+  public_key = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIExampleKeyMaterial deploy@example"
+}